// src/@v2/shared/domain/enum/security/origin-type.enum.ts
export enum OriginTypeEnum {
    HOMOGENEOUS_GROUP = 'HOMOGENEOUS_GROUP',

    DIRECTORY = 'DIRECTORY',
    MANAGEMENT = 'MANAGEMENT',
    SECTOR = 'SECTOR',
    SUB_SECTOR = 'SUB_SECTOR',
    OFFICE = 'OFFICE',
    SUB_OFFICE = 'SUB_OFFICE',

    WORKSTATION = 'WORKSTATION',
    EQUIPMENT = 'EQUIPMENT',
    ACTIVITIES = 'ACTIVITIES',
    GENERAL = 'GENERAL',
    SUPPORT = 'SUPPORT',
    OPERATION = 'OPERATION',
    ADMINISTRATIVE = 'ADMINISTRATIVE'
};